requirements.txt
asteroid-filterbanks >=0.4,<0.5
backports.cached_property
einops >=0.3,<0.4.0
hmmlearn >=0.2.7,<0.3
huggingface_hub >= 0.8.1
networkx >= 2.6,<3.0
omegaconf >=2.1,<3.0
pyannote.audio
pyannote.core >=4.4,<5.0
pyannote.database >=4.1.1,<5.0
pyannote.metrics >=3.2,<4.0
pyannote.pipeline >=2.3,<3.0
pydub
pytorch_lightning >=1.5.4,<1.7
pytorch_metric_learning >=1.0.0,<2.0
rich >= 12.0.0
semver >=2.10.2,<3.0
singledispatchmethod
soundfile >=0.10.2,<0.12
speechbrain >=0.5.12,<0.6
torch_audiomentations >= 0.11.0
torchmetrics >=0.11,<1.0
typing_extensions

python-dotenv

--extra-index-url https://download.pytorch.org/whl/cu116
torch
torchvision
torchaudio